+++ dcPower/include/global.h
@@ -17,9 +17,9 @@ char monitOut2[30];
 char monitOut3[30];
 char monitOut4[30];
 
-float under_volt_set;
-float nativeI_out;
-float nativeI_pri;
+double under_volt_set;
+double nativeI_out;
+double nativeI_pri;
 //--- PI control
 
 double ctrlError;
@@ -28,25 +28,25 @@ double preIntegral;
 double phaseShiftRatio;
 
 int code_run_input_select;
-float OC_Timer;
-float ovpTimer;
+double OC_Timer;
+double ovpTimer;
 
-float nativeIout;
-float nativeIpri;
-float nativeVout;
-float nativeVdc;
+double nativeIout;
+double nativeIpri;
+double nativeVout;
+double nativeVdc;
 
-float Pout;
-float Ipri;
-float Iout;
-float Vout;
+double Pout;
+double Ipri;
+double Iout;
+double Vout;
 
 int adcIpri;
 int adcIout;
 int adcVout;
 int adcVdc;
 
-float monitor[5]={0.0};
+double monitor[5]={0.0};
 int debug_count;
 int test_pulse_count;
 
@@ -61,32 +61,32 @@ int init_charge_flag = 0;
 int scib_rx_msg_flag = 0;
 int scic_rx_msg_flag = 0;
 
-float analog_ref_a;
-float analog_ref_b;
+double analog_ref_a;
+double analog_ref_b;
 //---
-float * dac_point[200];
-float dac_ch0_unit_count_value;
-float dac_ch1_unit_count_value;
-float dac_ch2_unit_count_value;
-float dac_ch3_unit_count_value;
+double * dac_point[200];
+double dac_ch0_unit_count_value;
+double dac_ch1_unit_count_value;
+double dac_ch2_unit_count_value;
+double dac_ch3_unit_count_value;
 
-float	inv_code_dac_scale_ch1;
-float	inv_code_dac_scale_ch2;
-float	inv_code_dac_scale_ch3;
-float	inv_code_dac_scale_ch4;
+double	inv_code_dac_scale_ch1;
+double	inv_code_dac_scale_ch2;
+double	inv_code_dac_scale_ch3;
+double	inv_code_dac_scale_ch4;
 //===================
 
 int dayRun ;
 int hourRun ;
 int secRun;
 
-float phaseVref;
-float VdcScale;
-float VdcOffset;
-float VoutScale;
-float VoutOffset;
+double phaseVref;
+double VdcScale;
+double VdcOffset;
+double VoutScale;
+double VoutOffset;
 
-float ref_time;
+double ref_time;
 CODE_INFO code_inform = {0,0,0,0,0,0,{0}};
 
 union PROTECT_FLAG protect_reg;
@@ -94,19 +94,19 @@ union PROTECT_FLAG protect_reg;
 int terminal_input_state;
 
 // PWM
-float	DutyRatio[3];
+double	DutyRatio[3];
 Uint16 DutyCount[3];
 
-float	Ts;
-float	inv_Ts;
-float 	gfRunTime=0;
+double	Ts;
+double	inv_Ts;
+double 	gfRunTime=0;
 int timer0_count;
-float	Vdc=0.0;
-float	Vs_max=0.0;
+double	Vdc=0.0;
+double	Vs_max=0.0;
 
-float refer_in0;
-float	reference_in=0.0;
-float	reference_out=0.0;
+double refer_in0;
+double	reference_in=0.0;
+double	reference_out=0.0;
 //--- Sci.c
 char gStr1[50] = {0};
 char gStr2[10] = {0};
@@ -129,33 +129,33 @@ int gRunFlag=0;
 int gTripSaveFlag=0;
 int EepromSaveFlag = 0;
 int Flag_TripDataBackup = 0;
-float OverCurLimit;
+double OverCurLimit;
 
-float codeVdcOffset;
-float codeVdcSpan;
-float codeVoutOffset;
-float codeVoutSpan;
+double codeVdcOffset;
+double codeVdcSpan;
+double codeVoutOffset;
+double codeVoutSpan;
 
-float code_Ki=50.0;
-float code_Kp=0.2;
+double code_Ki=50.0;
+double code_Kp=0.2;
 
-float code_integLimit=1.0;
-float code_phaseMax=0.7;
+double code_integLimit=1.0;
+double code_phaseMax=0.7;
 
-float codeVdcOffset;
-float codeVdcSpan;
-float codeVoutOffset;
-float codeVoutSpan;
+double codeVdcOffset;
+double codeVdcSpan;
+double codeVoutOffset;
+double codeVoutSpan;
 
-float code_adc_vdc_low;
-float code_adc_vdc_high;
-float code_Vdc_calc_low;
-float code_Vdc_calc_high;
+double code_adc_vdc_low;
+double code_adc_vdc_high;
+double code_Vdc_calc_low;
+double code_Vdc_calc_high;
 
-float code_adc_Vout_low;
-float code_adc_Vout_high;
-float code_Vout_calc_low;
-float code_Vout_calc_high;
+double code_adc_Vout_low;
+double code_adc_Vout_high;
+double code_Vout_calc_low;
+double code_Vout_calc_high;
 
 int code_protect_uv_off = 0;
 int code_protect_ov_off = 0;
@@ -167,28 +167,28 @@ int code_protect_ex_trip_off=0;
 //============================================
 //   CODE variable 
 //============================================
-float code_I_out_ref;        // 0
-float code_Vout;             // 1
-float code_VoutLimit;        // 2
-float code_ovpTime;          // 3
-float codeOcTime;           // 4
-float codePwmPhaseInit;      // 5
-float codeInitTime;          // 6
+double code_I_out_ref;        // 0
+double code_Vout;             // 1
+double code_VoutLimit;        // 2
+double code_ovpTime;          // 3
+double codeOcTime;           // 4
+double codePwmPhaseInit;      // 5
+double codeInitTime;          // 6
 
-float code_ctrl_mode;        // 10
-float codeSetPulseNumber;    // 11
-float code_testPwmPhase;     // 12
-float code_set_Vdc_on;       // 13
-float codeIoutAdcOffset;     // 14
-float codeIpriAdcOffset;     // 15
-float codeIoutScale;         // 16
-float codeIpriScale;         // 17
-float code_protect_inhibit_on;   // 18
+double code_ctrl_mode;        // 10
+double codeSetPulseNumber;    // 11
+double code_testPwmPhase;     // 12
+double code_set_Vdc_on;       // 13
+double codeIoutAdcOffset;     // 14
+double codeIpriAdcOffset;     // 15
+double codeIoutScale;         // 16
+double codeIpriScale;         // 17
+double code_protect_inhibit_on;   // 18
 
-float Data_Check;				// 800
-float Data_Backup;				// 801
-float Data_Load;				// 802
-float Data_Init;				// 803
+double Data_Check;				// 800
+double Data_Backup;				// 801
+double Data_Load;				// 802
+double Data_Init;				// 803
 //=====================================================
 #endif
